Refactor CEF app and browser process handler into separate files
