+++ src/cef/CMakeLists.txt
@@ -1,6 +1,8 @@
 set(cef_SOURCES
     $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/core/cef_init.cpp>
     $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/core/cef_init.h>
+    $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/core/cef_app.cpp>
+    $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/core/cef_app.h>
     $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/ui/uicefwebview.cpp>
     $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/ui/uicefwebview.h>
     $<$<BOOL:${USE_CEF}>:${CMAKE_CURRENT_LIST_DIR}/ui/cef_client.h>
